Refactors authentication and data fetching
Replaces axios with the native fetch API for authentication and data fetching to simplify the codebase and improve maintainability.

Introduces an image loader worker and meta cache service to optimize image loading and data caching in the artists and songs pages, improving performance and user experience. This includes caching mechanisms for both artists and song data, utilizing IndexedDB for persistence.

Handles 401 errors with a unified `fetchWithAuth` function which redirects to login page

Removes axios dependency.